chgore: renames files
- atomic => atomicObj
- atomicCss => atomic
