+++ __tests__/atomicObj.spec.js
@@ -1,0 +1,7 @@
+import { default as atomicObj } from '../src';
+
+describe('The atomic CSS object', () => {
+    it('should match the snapshot', () => {
+        expect(atomicObj).toMatchSnapshot();
+    });
+});
